user profile page divived into components
